kernels: every program id is read in 64-bit at its source (293 sites, 157 files) — the tile form of the wrap (group_norm at Mochi's VAE); the certifier stops at a sticky CUDA error (register 59)
Past add::14 the op-by-op run died at aten.native_group_norm::26:
`chan_start * HW` and `batch_idx * C * HW`, program-id-derived scalars
times dimensions — the same wrap in the tile form no arange regex could
see. Every `tl.program_id(k)` now reads `.to(tl.int64)` (FlagGems #6083's
form): every offset derived from a program id is 64-bit by construction.
The gate refuses an int32 program id (seen RED on an injected one). The
sort kernel's decoupled-lookback counter is cast back to int32 — a
loop-carried count keeps one type (26 sort cells failed to compile
before that cast; 84 pass after).

Beyond-2^31 group_norm test: N=1, C=64, HW=40e6 (the group at channel 56
starts at 2.24e9) — RED on 7ee3d7c (error 700), GREEN now; its first
form (HW=35.2e6, last group at 1.97e9) was green on the old tree and is
recorded as the vacuous form it was.

Measured on card 2, NBX_DISABLE_AUTOTUNE=1, before (7ee3d7c) vs after x3:
TinyLlama, Kokoro, Sana MultiLing (4 steps), whisper — byte-identical,
warm timings within 3 % (the first after run pays recompilation). Kernel
suite on card 3, pinned: 883 passed, 28 failed = 26 sort compile errors
(fixed above) + the 2 pinned-ordinal prefill cells (fail on 7ee3d7c under
the same pin, pass unpinned).

Certifier: after an illegal memory access every later launch and malloc
fails the same way; on 07:38 one such key cost 227 further keys reported
FAILED for "GPU malloc failed (error 700) for 256 bytes". The run now
stops at the first sticky error, names the key, exits 1 (3 tests,
injection seen RED 07:44).

+++ src/neurobrix/kernels/ops/add.py
@@ -11,7 +11,7 @@ def add_forward_kernel(
     BLOCK_SIZE: tl.constexpr,
 ):
     """out = x + alpha * y (tensor + tensor)"""
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
 
@@ -27,7 +27,7 @@ def add_scalar_kernel(
     BLOCK_SIZE: tl.constexpr,
 ):
     """out = x + scalar"""
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
 
@@ -57,7 +57,7 @@ def add_scalar_dev_kernel(
     exactly representable in f32. Every ATen call site passes alpha=±1
     (exact); a non-representable alpha would need an f64-carried arg.
     """
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
     s = (tl.load(s_ptr).to(tl.float64) * alpha).to(tl.float32)
@@ -82,7 +82,7 @@ def add_bias_broadcast_kernel(
     stride-0 broadcast but compatible with the kernel's flat 1D
     addressing.
     """
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     # Cast to int64 — tensors >= 2^31 elements (e.g. Sana 4Kpx VAE
     # add::88 input 1x4096x4096x128 = 2^31) overflow int32 offset
     # arithmetic silently and corrupt `offset % feat_dim`, producing
