+++ src/neurobrix/kernels/ops/pixel_unshuffle.py
@@ -44,7 +44,7 @@ def pixel_unshuffle_kernel(
 
     Each thread handles one output element.
     """
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     idx = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = idx < n_elements
